Autobuild: Use actions/cache for external deps
This commit introduces an actions/cache step for Android, Mac and Windows
builds. For Linux, this makes little sense as all dependencies are installed
via apt, which should be almost locally anyway (Azure apt mirror).

We rely on pinned versions for the relevant external dependencies.

The following environments are cached:
- Qt installation directory (Windows, Mac, Android)
- pip cache (Windows, Mac)
- choco cache (Windows)
- Android SDK & NDK (Android)

The cache is used if the workflow-provided calls (which include Qt
versions) are unchanged and if certain files (most notably the files
that this commit touches) are unchanged.
If anything changes, the cache is not used.
Instead, everything starts clean and the result is cached for further
reuse again.

This is supposed to:
- Make builds more deterministic
- Speed up builds by avoiding repeated downloads and some extractions
- Increase resilience against temporary external outages (e.g. Qt mirror
  problems)