Country flags: Use QLocale::matchingLocales in place of hardcoded table
The country-code-to-flag-image handling has traditionally used a rather
large switch/case statement to retain support for Qt4 (according to the
comments). We no longer support building with Qt4.

A Qt-native approach had already been developed by @corrados, but it was
disabled. This commit removes the switch-based lookup and enables the
QLocale::matchingLocales-based lookup instead.

It has been verified that the new approach handles all of the
previously supported codes.
The new code adds support for all codes which are available in the
current Qt version and where flag icons are available. When building on
Qt 5.15.2 or 6.2.3 this results in support for 53 new countries/flags
compared to the old logic. Those countries become selectable in clients'
user profile settings and in servers' country settings.
Clients which are based on the old logic will show a placeholder if they
can't resolve an unknown country code.

Co-authored-by: Volker Fischer <corrados@users.noreply.github.com>