Android: Set ANDROID_VERSION_{CODE,NAME}
android/AndroidManifest.xml uses specific macros
(%%INSERT_VERSION...). Those macros are replaced by the
ANDROID_VERSION_* qmake variables, which were previously unset and
therefore defaulted to "1". This made it impossible to update the
APK without uninstalling first.

This commit populates these variables. The user-visible
ANDROID_VERSION_NAME is set similar to VERSION for other platforms.
The Android-internal ANDROID_VERSION_CODE is set to the number of git
commits in the current branch. This seems to be the most straightforward
solution which guarantees unique, increasing integers within the scope
of the master branch. Feature branches will break this logic, so
developers may still be required to uninstall/reinstall.

In order to get the commit count, the autobuild Android logic had to be modified
to fetch the repo in full depth instead of only the most-recent commit.

Tested via `apktool d` and by checking `apktool.yml`.

Related: #1760