Autobuild: Simplify variable outputs in analyze_git_references.py
The following variable is not referenced in the Autobuild workflow at
all and can therefore be safely dropped from the script output:
PUSHED_NAME

The following variables are referenced in the Autobuild workflow and are
declared as outputs, but those outputs are not referred to anywhere.
Therefore, they can safely be dropped:
- PUSHED_NAME
- X_GITHUB_WORKSPACE

The following variables are only relevant for releases
(PUBLISH_TO_RELEASE=true):
- RELEASE_TITLE
- RELEASE_TAG
- IS_PRERELEASE

Therefore, we only need to calculate their value for actual releases.

This vastly simplifies the script logic.