Android: Detect liboboe sources automatically
Instead of hardcoding the list of relevant .cpp and .h files, use
qmake's $$files to detect the sources. This makes liboboe updates
easier and makes Jamulus.pro more readable.

It has been confirmed that the hardcoded list of .cpp and .h files
closely matches the list of available files. In other words, this commit does
not change the content of the variables a lot, it just switches from
hardcoding to dynamic detection.
The new approach adds a single new .h file which was not listed
previously:
libs/oboe/src/flowgraph/resampler/KaiserWindow.h