Autobuild: Merge Mac build scripts
- `diff -u autobuild/mac/{artifacts,codeQL}/autobuild_mac_1_prepare.sh` is
  empty.

- `diff -u autobuild/mac/{artifacts,codeQL}/autobuild_mac_2_build.sh`
  contains extra logic for signing, but as this is controlled by a
  parameter, it is fine to merge anyway.

- `autobuild/mac/artifacts/autobuild_mac_3_copy_files.sh` does not exist
  for the CodeQL job (it's unused there).

Therefore, it's safe to move those scripts one level up and drop all the
duplication. That makes it consistent with all the other platforms as
well.

This commit moves the .sh files to their new location and adjusts a
single line to account for the reduced nesting level. It also updates
the callers to use the new paths.