Jamulus.pro: Drop bad QMAKE_APPLICATION_BUNDLE_NAME.
Jamulus.pro sets `QMAKE_APPLICATION_BUNDLE_NAME.` (note the trailing
dot).
This sets the variable with exactly this name, but not the probably
intended dot-less QMAKE_APPLICATION_BUNDLE_NAME.

As this can't have worked in the past, it's obviously not necessary. So
instead of fixing the assignment, the line is dropped altogether.

The docs[1] suggest that QMAKE_APPLICATION_BUNDLE_NAME is used in the
plist generation logic, but that the relevant place defaults to using
TARGET, which we do set properly. This highlights that it should be ok
to work without QMAKE_APPLICATION_BUNDLE_NAME.

[1] https://doc.qt.io/qt-5/qmake-variable-reference.html